+++ src/main/java/com/project/Haru_Mail/common/config/WebConfig.java
@@ -1,4 +1,4 @@
-package com.project.Haru_Mail.common;
+package com.project.Haru_Mail.common.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
@@ -12,7 +12,7 @@ public class WebConfig implements WebMvcConfigurer {
     public void addCorsMappings(CorsRegistry registry) {
         registry.addMapping("/**")  // 모든 경로에 대해 CORS 설정
                 .allowedOrigins("http://localhost:5173/")  // 클라이언트 도메인 허용
-                .allowedMethods("GET", "POST", "PUT", "DELETE", "OPTIONS")  // 허용할 HTTP 메서드
+                .allowedMethods("*")  // 허용할 HTTP 메서드
                 .allowedHeaders("*")  // 모든 헤더 허용
                 .allowCredentials(true);  // 쿠키를 포함한 요청 허용
     }
